requirements.txt
click==8.0.3
Django==4.1.2
jellyfish==0.9.0
pandas==1.5.1
python_magic==0.4.27
